requirements.txt
autopep8==1.4.4
Django==2.2.7
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
